client/requirements.txt
sounddevice>=0.4.6
scipy>=1.11.0
pynput>=1.7.6
pyautogui>=0.9.54
numpy>=1.24.0
openai-whisper>=20231117
pystray>=0.19.5
Pillow>=10.0.0
cairosvg>=2.7.0
